slight cleanup of player.js code
changed if block (when only one var could be true) to a switch.

apparently a switch if faster than an if else chain in js, so thats nice.

added var names for keyTypes and skipvalue to remove magic numbers and so that i can easily change the value later on should i want to